+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* server/cluster: Improve cluster performance
 * tagger/quote: Don't mis-tag starts of chapters as suspensions
 
 ## 2.0.0-beta5 (2018-12-07)